requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.2
opencv-python-headless==4.9.0.80
